use DDD package and class structure
