+++ src/main/java/hu/adam/sipos/music/artist/ArtistDto.java
@@ -1,6 +1,7 @@
-package hu.adam.sipos.music.dtos;
+package hu.adam.sipos.music.artist;
 
-import hu.adam.sipos.music.domain.Artist;
+import hu.adam.sipos.music.album.AlbumDto;
+import hu.adam.sipos.music.artist.Artist;
 
 import java.util.ArrayList;
 import java.util.List;
