+++ src/main/java/hu/adam/sipos/music/artist/ArtistFileRepository.java
@@ -1,8 +1,6 @@
-package hu.adam.sipos.music.persistence;
+package hu.adam.sipos.music.artist;
 
-import hu.adam.sipos.music.domain.Artist;
 import hu.adam.sipos.music.domain.repository.AbstractRepository;
-import hu.adam.sipos.music.dtos.ArtistDto;
 import hu.adam.sipos.music.serialization.SerializationService;
 
 import java.io.IOException;
