+++ src/main/java/hu/adam/sipos/music/library/Library.java
@@ -1,10 +1,16 @@
-package hu.adam.sipos.music.domain;
+package hu.adam.sipos.music.library;
+
+import hu.adam.sipos.music.album.Album;
+import hu.adam.sipos.music.artist.Artist;
+import hu.adam.sipos.music.playlist.Playlist;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.OptionalInt;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class Library {
     public List<Artist> artists = new ArrayList<>();
@@ -22,6 +28,24 @@ public class Library {
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
+    }
+
+    public Library addOrOverride(Artist artist) {
+        if (artist == null) {
+            return this;
+        }
+
+        OptionalInt artistIndexInLibraryOptional = IntStream.range(0, artists.size())
+                .filter(i -> artists.get(i).getName().equals(artist.getName()))
+                .findFirst();
+
+        if (artistIndexInLibraryOptional.isPresent()) {
+            artists.set(artistIndexInLibraryOptional.getAsInt(), artist);
+        } else {
+            artists.add(artist);
+        }
+
+        return this;
     }
 
     public List<Artist> getArtists() {
