+++ src/main/java/hu/adam/sipos/music/library/LibraryFileRepository.java
@@ -1,8 +1,8 @@
-package hu.adam.sipos.music.persistence;
+package hu.adam.sipos.music.library;
 
-import hu.adam.sipos.music.domain.Library;
+import hu.adam.sipos.music.library.Library;
 import hu.adam.sipos.music.domain.repository.AbstractRepository;
-import hu.adam.sipos.music.dtos.LibraryDto;
+import hu.adam.sipos.music.library.LibraryDto;
 import hu.adam.sipos.music.serialization.SerializationService;
 
 import java.io.IOException;
