+++ src/test/java/hu/adam/sipos/music/album/AlbumTest.java
@@ -1,5 +1,8 @@
-package hu.adam.sipos.music.domain;
+package hu.adam.sipos.music.album;
 
+import hu.adam.sipos.music.album.Album;
+import hu.adam.sipos.music.artist.Artist;
+import hu.adam.sipos.music.track.Track;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
