+++ src/test/java/hu/adam/sipos/music/playlist/PlaylistTest.java
@@ -1,5 +1,8 @@
-package hu.adam.sipos.music.domain;
+package hu.adam.sipos.music.playlist;
 
+import hu.adam.sipos.music.library.Library;
+import hu.adam.sipos.music.playlist.Playlist;
+import hu.adam.sipos.music.track.Track;
 import org.junit.Assert;
 import org.junit.Test;
