+++ src/test/java/hu/adam/sipos/music/serialization/JsonSerializationServiceTest.java
@@ -1,10 +1,10 @@
 package hu.adam.sipos.music.serialization;
 
-import hu.adam.sipos.music.domain.Album;
-import hu.adam.sipos.music.domain.Artist;
-import hu.adam.sipos.music.domain.Library;
-import hu.adam.sipos.music.domain.Track;
-import hu.adam.sipos.music.dtos.LibraryDto;
+import hu.adam.sipos.music.album.Album;
+import hu.adam.sipos.music.artist.Artist;
+import hu.adam.sipos.music.library.Library;
+import hu.adam.sipos.music.track.Track;
+import hu.adam.sipos.music.library.LibraryDto;
 import org.junit.Assert;
 import org.junit.Test;
